Simple implementation of scanf and printf
This version is compiled using a 32-bit computer. The method of function parameter passing is implemented by stack. If you know the whole process of C calling function in 32-bit computer, you will understand this code. If you use GCC to compile and run this code, an error will be reported. If you want to run this code, you can use VC6.0 and visual studio to choose the default 32-bit compilation method to compile and run this code